Make UserSortable a proper concern
It confuses the heck out of YARD + aligns it with other model concerns +
removed the need to disable the "1 class per file" rule of Rubocop
